cleanup: Consolidate  onboarding code into platform/cloud/onboarding
