+++ src/platform/cloud/onboarding/CloudAuthTimeoutView.vue
