+++ src/platform/cloud/onboarding/CloudForgotPasswordView.vue
