+++ src/platform/cloud/onboarding/CloudSignupView.vue
