+++ src/platform/cloud/onboarding/assets/css/fonts.css
