+++ src/platform/cloud/onboarding/assets/videos/thumbnail.png
