+++ src/platform/cloud/onboarding/auth.ts
