+++ src/platform/cloud/onboarding/components/CloudLayoutView.vue
