+++ src/platform/cloud/onboarding/components/CloudLogo.vue
