+++ src/platform/cloud/onboarding/components/CloudSignInForm.vue
