+++ src/platform/cloud/onboarding/components/CloudTemplateFooter.vue
