+++ src/platform/cloud/onboarding/skeletons/CloudLoginViewSkeleton.vue
