+++ src/platform/cloud/onboarding/skeletons/CloudSurveyViewSkeleton.vue
